+++ .env
@@ -1,3 +1,2 @@
 MONGO_URI=
 DB_NAME=
-JWT_SECRET=dwdwdwdwdwdw
