+++ package.json
@@ -5,15 +5,12 @@
   "license": "MIT",
   "dependencies": {
     "@types/jest": "^29.5.11",
-    "@types/jsonwebtoken": "^9.0.5",
     "@types/supertest": "^6.0.2",
     "bcrypt": "^5.1.1",
     "dotenv": "^16.3.1",
     "express": "^4.18.2",
     "express-validator": "^7.0.1",
     "jest": "^29.7.0",
-    "jsonwebtoken": "^9.0.2",
-    "jwt": "^0.2.0",
     "mongodb": "^6.3.0",
     "ngrok": "^5.0.0-beta.2",
     "supertest": "^6.3.3",
