+++ src/app.ts
@@ -21,6 +21,4 @@ app.use('/users', userRoute)
 
 app.use('/auth', authRoute)
 
-// app.use('/comments', commentsRouter)
-
 app.use('/testing/all-data', deleteAllDataRoute)
