+++ src/db/db.ts
@@ -2,12 +2,12 @@ import dotenv from 'dotenv'
 import {MongoClient} from "mongodb";
 import {BlogDb} from "../models/blogs/db/blog-db";
 import {PostDbType} from "../models/posts/db/post-db";
+import any = jasmine.any;
 import {UserDbType} from "../models/users/db/user-db";
-import {CommentInputType} from "../models/comments/input/comment-input";
 
 dotenv.config()
 // указываем порт
-export const port = process.env.PORT || 3010
+export const port = process.env.PORT || 3007
 
 // указываем ссылку для коннекта к базе
 // локальная - mongodb://localhost:27017
@@ -24,8 +24,6 @@ export const blogsCollection = dataBase.collection<BlogDb>('blogs')
 export const postsCollection = dataBase.collection<PostDbType>('posts')
 
 export const usersCollection = dataBase.collection<UserDbType>('users')
-
-export const commentsCollection = dataBase.collection<CommentInputType>('users')
 
 export const runDb = async () => {
     try {
