--- src/models/comments/commentator-info/commentator-info.ts
@@ -1,4 +1,0 @@
-export type CommentatorInfo = {
-    userId: string
-    userLogin: string
-}
