--- src/models/comments/input/comment-input.ts
@@ -1,8 +1,0 @@
-import {CommentatorInfo} from "../commentator-info/commentator-info";
-
-export type CommentInputType = {
-    content: string
-    commentatorInfo: CommentatorInfo
-    postId: string
-    createdAt: string
-}
