--- src/models/comments/output/comment-output.ts
@@ -1,9 +1,0 @@
-export type CommentOutputType = {
-    id: string,
-    content: string,
-    commentatorInfo: {
-        userId: string,
-        userLogin: string
-    },
-    createdAt: string
-}
