+++ src/models/common.ts
@@ -4,7 +4,7 @@ export type ParamType = {
     id: string
 }
 
-// export type CustomRequest<ResBody = any> = Request & {userId: any}
+export type RequestType = Request<{}, {}, {}, {}>
 
 export type RequestWithParams<P> = Request<P, {}, {}, {}>
 
@@ -17,8 +17,6 @@ export type RequestWithParamsAndQuery<P, Q> = Request<P, {}, {}, Q>
 export type RequestWithParamsAndBody<P, B> = Request<P, {}, B, {}>
 
 export type ResponseType<T> = Response<T, {}>
-
-
 
 export const HTTP_RESPONSE_CODES = {
     SUCCESS: 200,
