+++ src/repositories/user-query-repository.ts
@@ -1,10 +1,18 @@
-import {usersCollection} from "../db/db";
+import {postsCollection, usersCollection} from "../db/db";
+import {postMapper} from "../models/posts/mappers/post-mapper";
+import {OutputPostModel} from "../models/posts/output/output-post";
 import {ObjectId, SortDirection, WithId} from "mongodb";
-import {PaginationType} from "../models/common";
+import {PaginationType, ResponseType} from "../models/common";
 import {QueryUserInputModel} from "../models/users/input/query.user.input.model";
 import {UserDbType} from "../models/users/db/user-db";
 import {OutputUser} from "../models/users/output/output-user";
-import {CommentatorInfo} from "../models/comments/commentator-info/commentator-info";
+
+type NewPostDataType = {
+    title: string,
+    shortDescription: string,
+    content: string,
+    blogId: string
+}
 
 export type QueryPostDataType = {
     pageNumber: number
@@ -75,29 +83,5 @@ export class UserQueryRepository {
             email: users.email,
             login: users.login
         }
-    }
-
-    static async getCommentatorById(commentatorId: string): Promise<CommentatorInfo | null> {
-        try {
-
-            const commentator = await usersCollection.findOne({_id: new ObjectId(commentatorId)})
-
-            if (!commentator) {
-                return null
-            } else {
-                return {
-                    userId: commentator._id.toString(),
-                    userLogin: commentator.login
-                }
-            }
-
-
-        } catch (e) {
-
-            console.error(e)
-
-            return null
-        }
-
     }
 }
