--- src/routes/comments-route
@@ -1,3 +1,0 @@
-import express, {Request, Response} from 'express'
-
-export const commentsRouter = express.Router()
