--- src/types/index.d.ts
@@ -1,5 +1,0 @@
-export declare module 'express-serve-static-core' {
-    interface Request {
-        userId: string
-    }
-}
